+++ Mon-Ami.Server/Mon-Ami.API/Controllers/ActivitiesController.cs
@@ -23,26 +23,26 @@ namespace Mon_Ami.API.Controllers
         [HttpGet]
         public async Task<ActionResult<List<Activity>>> GetActivityList(CancellationToken token)
         {
-            List<Activity> result = await _mediator.Send(new ActivityList.Query(), token);
+            List<Activity> result = await _mediator.Send(new List.Query(), token);
             return result;
         }
 
         [HttpGet("{id}")]
         public async Task<ActionResult<Activity>> GetActivity(Guid id, CancellationToken token)
         {
-            Activity result = await _mediator.Send(new ActivityDetails.Query { Id = id }, token);
+            Activity result = await _mediator.Send(new Details.Query { Id = id }, token);
             return result;
         }
 
         [HttpPost]
-        public async Task<ActionResult<Unit>> CreateActivity(ActivityCreate.Command command, CancellationToken token)
+        public async Task<ActionResult<Unit>> CreateActivity(Create.Command command, CancellationToken token)
         {
             Unit result = await _mediator.Send(command, token);
             return result;
         }
 
         [HttpPut("{id}")]
-        public async Task<ActionResult<Unit>> EditActivity(Guid id, ActivityEdit.Command command, CancellationToken token)
+        public async Task<ActionResult<Unit>> EditActivity(Guid id, Edit.Command command, CancellationToken token)
         {
             command.Id = id;
             Unit result = await _mediator.Send(command, token);
@@ -50,10 +50,10 @@ namespace Mon_Ami.API.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<ActionResult<Unit>> DeleteActivity(Guid id, ActivityDelete.Command command, CancellationToken token)
+        public async Task<ActionResult<Unit>> DeleteActivity(Guid id, Delete.Command command, CancellationToken token)
         {
             command.Id = id;
-            Unit result = await _mediator.Send(new ActivityDelete.Command { Id = id }, token);
+            Unit result = await _mediator.Send(new Delete.Command { Id = id }, token);
             return result;
         }
     }
